Don't delete example.png >:(

+++ js/util/plot.js
@@ -74,6 +74,7 @@ function getChart(tag, msg) {
 async function deleteCharts() {
 	const dir = await readdir('./img/');
 	dir.forEach((file) => {
+		if(file === "example.png") return;
 		fs.unlink(`img/${file}`, (err) => {
 			if (err) throw err;
 		});
